+++ v2installer/process/prelimconfig.bash
@@ -1,7 +1,7 @@
 # Part of Alaterm, version 2.
 # Routines for preliminary configuration, after unpacking archive.
 
-echo "$(caller)" | grep -F alaterm-installer >/dev/null 2>&1
+echo "$(caller)" | grep -e alaterm-installer >/dev/null 2>&1
 if [ "$?" -ne 0 ] ; then
         echo "This file is not stand-alone."
         echo "It must be sourced from alaterm-installer."
@@ -10,6 +10,7 @@ fi
 
 
 cd "$alatermTop/etc"
+touch pacman.conf
 sed -i '/^#Color/s/^#//' pacman.conf 2>/dev/null
 chmod 644 pacman.conf
 
@@ -21,16 +22,20 @@ install_template "bash.bashrc.bash" # Does nothing.
 install_template "prelim-profile.bash" # As temporary /etc/profile.
 # Create a temporary command for launching Arch as root:
 install_template "prelim-launch.bash" "755"
-# Now execute that command, which activates /root.bashrc:
-bash "$alatermTop/prelim-launch.bash"
+# Now execute that command, which activates /root.bashrc, except in devmode:
+if [ -z "$devmode" ] ; then
+	bash "$alatermTop/prelim-launch.bash"
+else
+	gotSudo="yes" && echo "gotSudo=yes" >> "$alatermTop/status"
+fi
 temprl="$(grep gotSudo $alatermTop/status 2>/dev/null)"
 if [ "$temprl" = "" ] ; then # Failed prelim-launch.bash.
 	echo -e "$PROBLEM Could not install sudo."
 	rm -f "$alatermTop/prelim-launch.bash"
 	exit 1
 fi
-rm -f "$alatermTop/prelim-launch.bash"
-echo -e "$INFO Preliminary configuration done."
-echo "completedPrelim=yes" >> "$alatermTop/status"
+[ -z "$devmode" ] && rm -f "$alatermTop/prelim-launch.bash"
+[ -z "$devmode" ] && echo -e "$INFO Preliminary configuration done."
+completedPrelim="yes" && echo "completedPrelim=yes" >> "$alatermTop/status"
 sleep .2
 ##
